backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.1
PyYAML==6.0
djangorestframework-simplejwt==4.7.2
django-filter==23.1